AuthenticationDto and LoginController moved to security/jwt/model and controller
